[fix] correções solicitadas

+++ event-manager-front-end/src/app/features/views/event-components/event-components.component.ts
@@ -1,11 +1,25 @@
 import { Component, OnInit } from '@angular/core';
-import { FormControl, FormGroup, FormsModule, Validators } from '@angular/forms';
+import {
+  FormControl,
+  FormGroup,
+  FormsModule,
+  Validators,
+} from '@angular/forms';
 import { CommonModule } from '@angular/common';
-import { FormGroupArray, IOptions } from '../../../core/interface/form.interface';
+import {
+  FormGroupArray,
+  IOptions,
+} from '../../../core/interface/form.interface';
 import { FormFieldEnum } from '../../../core/enums/formFieldEnum';
-import { FormComponent } from "../../../core/components/form-group/form/form.component";
+import { FormComponent } from '../../../core/components/form-group/form/form.component';
 import { ApiService } from '../../services/api.service';
-import { IRequestEmployee, IRequestEmployeeType, IRequestMaterial, IResponseEmployee, IResponseMaterial } from '../../../core/interface/event.interface';
+import {
+  IRequestEmployee,
+  IRequestEmployeeType,
+  IRequestMaterial,
+  IResponseEmployee,
+  IResponseMaterial,
+} from '../../../core/interface/event.interface';
 import { MatButtonModule } from '@angular/material/button';
 import { OptionsService } from '../../services/options.service';
 import { MatPaginatorModule, PageEvent } from '@angular/material/paginator';
@@ -13,13 +27,13 @@ import { MatDialog } from '@angular/material/dialog';
 import { ModalUpdateGarcomComponent } from '../../../core/components/modais/modal-update-garcom/modal-update-garcom.component';
 import ModalUpdateMaterialComponent from '../../../core/components/modais/modal-update-material/modal-update-material.component';
 import { MaskEnum } from '../../../core/enums/maskEnum';
-import { MatMenuModule } from "@angular/material/menu";
+import { MatMenuModule } from '@angular/material/menu';
 import { MatIconModule } from '@angular/material/icon';
 import { ModalViewPartyWaiterComponent } from '../../../core/components/modais/modal-view-party-waiter/modal-view-party-waiter.component';
 import { BehaviorSubject } from 'rxjs';
 import { ToastService } from '../../../core/services/toast.service';
-
-
+import { ModalDeleteConfirmationComponent } from '../../../core/components/modais/modal-delete-confirmation/modal-delete-confirmation.component';
+import { ModalPixTelConfirmationComponent } from '../../../core/components/modais/modal-pix-tel-confirmation/modal-pix-tel-confirmation.component';
 
 @Component({
   selector: 'app-event-components',
@@ -32,20 +46,16 @@ import { ToastService } from '../../../core/services/toast.service';
     MatPaginatorModule,
     MatMenuModule,
     MatIconModule,
-],
+  ],
   templateUrl: './event-components.component.html',
-  styleUrl: './event-components.component.scss'
+  styleUrl: './event-components.component.scss',
 })
-export class EventComponentsComponent implements OnInit{
+export class EventComponentsComponent implements OnInit {
   public formPrimary = new FormGroup({
-    name: new FormControl<string | null>(null),
-    pixKey: new FormControl<string | null>(null),
-    phone: new FormControl<string | null>(null),
-    statusEmployee: new FormControl<string | null>(null),
-  });
-
-  public formSecondary = new FormGroup({
-     idEmployeeType: new FormControl<number | null>(null),
+    name: new FormControl<string | null>(null, [Validators.required]),
+    pixKey: new FormControl<string | null>(null, [Validators.required]),
+    phone: new FormControl<string | null>(null, [Validators.required]),
+    idEmployeeType: new FormControl<number | null>(null, [Validators.required]),
   });
 
   public formThird = new FormGroup({
@@ -63,22 +73,42 @@ export class EventComponentsComponent implements OnInit{
   public totalElements = 0;
   public createEmployeeType: boolean = false;
   public employeeTypesOptions$ = new BehaviorSubject<IOptions[]>([]);
-  
+
+  public formGroupItensPrimary: FormGroupArray = [];
 
   constructor(
     private readonly _service: ApiService,
     private readonly _optionsService: OptionsService,
     private readonly _dialog: MatDialog,
     private readonly _toast: ToastService
-  ) { }
-
-  ngOnInit(): void {
-     this.getEmployee();
-     this.getMaterials();
-     this.getEmployeeTypesOptions();
+  ) {
+    this.formGroupItensPrimary = this._formGroupItensPrimary();
   }
 
-  public get formGroupItensPrimary(): FormGroupArray{
+  ngOnInit(): void {
+    this.getEmployee();
+    this.getMaterials();
+    this.getEmployeeTypesOptions();
+    this.formPrimary.controls.phone.valueChanges.subscribe((value) => {
+      if (value && value.length >= 11) {
+        this._dialog.open(ModalPixTelConfirmationComponent, {
+          width: '400px',
+          height: '200px',
+          autoFocus: false,
+        }).afterClosed().subscribe((confirmed: boolean) => {
+          if (confirmed) {
+            this.formPrimary.controls.pixKey.setValue(value);
+          } else{
+            const pix = this.formGroupItensPrimary.findIndex((item => item.controlName === 'pixKey'));
+            this.formGroupItensPrimary[pix].hidden = false;
+            this.formPrimary.controls.pixKey.setValue(null);
+          } 
+        });
+      }
+    });
+  }
+
+  private _formGroupItensPrimary(): FormGroupArray {
     return [
       {
         component: FormFieldEnum.INPUT,
@@ -88,17 +118,6 @@ export class EventComponentsComponent implements OnInit{
         placeholder: 'Digite seu nome',
         size: '6',
         maxlength: 50,
-        mask: MaskEnum.NOME
-      },
-      {
-        component: FormFieldEnum.INPUT,
-        label: 'Chave Pix',
-        controlName: 'pixKey',
-        type: 'text',
-        placeholder: 'Digite sua chave pix',
-        size: '6',
-        mask: MaskEnum.PIX,
-        maxlength: 17
       },
       {
         component: FormFieldEnum.INPUT,
@@ -108,34 +127,31 @@ export class EventComponentsComponent implements OnInit{
         placeholder: 'Digite seu telefone',
         size: '6',
         mask: MaskEnum.PHONE,
-        maxlength: 15
+        maxlength: 15,
       },
-      {
-        component: FormFieldEnum.SELECT,
-        label: 'Status',
-        controlName: 'statusEmployee',
-        type: 'select',
-        options: this._optionsService.getOptionsStatusEmployee(),
-        size: '6',
-      },
-    ]
-  }
-
-  public get formGroupItensSecondary(): FormGroupArray{
-    return [
       {
         component: FormFieldEnum.SELECT,
         label: 'Tipo de Funcionário',
         controlName: 'idEmployeeType',
         type: 'select',
         options: this.employeeTypesOptions$.asObservable(),
-        size: '12',
-        
-      }
-    ]
+        size: '6',
+      },
+      {
+        component: FormFieldEnum.INPUT,
+        label: 'Chave Pix',
+        controlName: 'pixKey',
+        type: 'text',
+        placeholder: 'Digite sua chave pix',
+        size: '6',
+        mask: MaskEnum.PIX,
+        maxlength: 17,
+        hidden: true,
+      },
+    ];
   }
 
-  public get formGroupItensThird(): FormGroupArray{
+  public get formGroupItensThird(): FormGroupArray {
     return [
       {
         component: FormFieldEnum.INPUT,
@@ -143,42 +159,41 @@ export class EventComponentsComponent implements OnInit{
         controlName: 'description',
         type: 'text',
         size: '12',
-      }
-    ]
+      },
+    ];
   }
 
-  public get formGroupItensEmployeeType(): FormGroupArray{
+  public get formGroupItensEmployeeType(): FormGroupArray {
     return [
       {
         component: FormFieldEnum.INPUT,
-        label: 'Tipo de Funcionário',
+        label: 'Novo Tipo de Funcionário',
         controlName: 'type',
         type: 'text',
         size: '12',
-      }
-    ]
+      },
+    ];
   }
 
-  public createEmployee(){
+  public createEmployee() {
     const employee = this.formPrimary.value;
     const data = {
       ...employee,
-      idEmployeeType: this.formSecondary.value.idEmployeeType
+      idEmployeeType: this.formPrimary.value.idEmployeeType,
     } as IRequestEmployee;
 
-    this._service.postCreateEmployee(data).subscribe(({
-      next:(value) => {
+    this._service.postCreateEmployee(data).subscribe({
+      next: (value) => {
         this.formPrimary.reset();
-        this.formSecondary.reset();
         this.getEmployee();
+        const pix = this.formGroupItensPrimary.findIndex((item) => item.controlName === 'pixKey');
+        this.formGroupItensPrimary[pix].hidden = true;
         this._toast.success('Funcionário criado com sucesso!');
       },
       error: (err) => {
         this._toast.error('Erro ao criar funcionário. Tente novamente.', err);
-      }
-    }) 
-      
-    );
+      },
+    });
   }
 
   public createEmployeeTypeRequest() {
@@ -203,11 +218,13 @@ export class EventComponentsComponent implements OnInit{
   }
 
   public getEmployee() {
-    this._service.getEmployees(this.page, this.pageSize).subscribe((response) => {
-      this.ListEmployees = response.content;
-      this.totalElements = response.page.totalElements;
-      this.pageSize = response.page.size;
-    });
+    this._service
+      .getEmployees(this.page, this.pageSize)
+      .subscribe((response) => {
+        this.ListEmployees = response.content;
+        this.totalElements = response.page.totalElements;
+        this.pageSize = response.page.size;
+      });
   }
   public onChangePage(event: PageEvent) {
     this.page = event.pageIndex;
@@ -220,12 +237,12 @@ export class EventComponentsComponent implements OnInit{
     });
   }
 
-  public editEmployee(id: number){
+  public editEmployee(id: number) {
     const dialogRef = this._dialog.open(ModalUpdateGarcomComponent, {
       width: '600px',
       height: '600px',
       data: {
-        id: id
+        id: id,
       },
     });
     dialogRef.afterClosed().subscribe((result) => {
@@ -235,7 +252,7 @@ export class EventComponentsComponent implements OnInit{
     });
   }
 
-  public viewFestas(id: number){
+  public viewFestas(id: number) {
     const dialogRef = this._dialog.open(ModalViewPartyWaiterComponent, {
       width: '60vw',
       height: '70vh',
@@ -243,7 +260,7 @@ export class EventComponentsComponent implements OnInit{
       maxHeight: '100vh',
       autoFocus: false,
       data: {
-        id: id
+        id: id,
       },
     });
     dialogRef.afterClosed().subscribe((result) => {
@@ -254,11 +271,19 @@ export class EventComponentsComponent implements OnInit{
   }
 
   public deleteEmployee(id: number) {
-    this._service.deleteEmployee(id).subscribe(() => {
-      this.getEmployee();
+    const dialog = this._dialog.open(ModalDeleteConfirmationComponent, {
+      width: '400px',
+      height: '200px',
+      autoFocus: false,
+    });
+    dialog.afterClosed().subscribe((result) => {
+      if (result) {
+        this._service.deleteEmployee(id).subscribe(() => {
+          this.getEmployee();
+        });
+      }
     });
   }
-
   public createMaterial() {
     const material: IRequestMaterial = this.formThird.value as IRequestMaterial;
     this._service.postCreateMaterial(material).subscribe(() => {
@@ -268,8 +293,17 @@ export class EventComponentsComponent implements OnInit{
   }
 
   public deleteMaterial(id: number) {
-    this._service.deleteMaterial(id).subscribe(() => {
-      this.getMaterials();
+    const dialog = this._dialog.open(ModalDeleteConfirmationComponent, {
+      width: '400px',
+      height: '200px',
+      autoFocus: false,
+    });
+    dialog.afterClosed().subscribe((result) => {
+      if (result) {
+        this._service.deleteMaterial(id).subscribe(() => {
+          this.getMaterials();
+        });
+      }
     });
   }
 
@@ -278,7 +312,7 @@ export class EventComponentsComponent implements OnInit{
       width: '400px',
       height: '300px',
       data: {
-        id: id
+        id: id,
       },
     });
     dialogRef.afterClosed().subscribe((result) => {
@@ -288,4 +322,3 @@ export class EventComponentsComponent implements OnInit{
     });
   }
 }
-
